pyproject.toml
[project]
name = "home-ai-os"
version = "0.1.0"
description = "Unified Personal / Family / Project Operations System"
requires-python = ">=3.12"
license = {text = "MIT"}

dependencies = [
    # Web Framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.9",

    # Database
    "sqlalchemy[asyncio]>=2.0.30",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.13.0",
    "pgvector>=0.3.0",

    # Redis & Celery
    "redis>=5.0.0",
    "celery[redis]>=5.4.0",

    # Validation & Settings
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",

    # Telegram Bot
    "aiogram>=3.10.0",

    # AI - Anthropic
    "anthropic>=0.34.0",

    # AI - Ollama (local models)
    "ollama>=0.3.0",

    # AI - Transcription
    "faster-whisper>=1.0.0",

    # AI - Embeddings
    "sentence-transformers>=3.0.0",

    # OCR
    "pytesseract>=0.3.10",
    # easyocr — install separately, heavy deps

    # S3
    "boto3>=1.34.0",

    # Document processing
    "python-docx>=1.1.0",
    "pypdf>=4.0.0",
    "openpyxl>=3.1.0",

    # Media processing
    "Pillow>=10.0.0",

    # Auth
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",

    # Logging
    "structlog>=24.0.0",

    # HTTP client
    "httpx>=0.27.0",

    # Utilities
    "python-dateutil>=2.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "factory-boy>=3.3.0",
    "httpx>=0.27.0",
    "ruff>=0.5.0",
    "mypy>=1.10.0",
    "pre-commit>=3.7.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "workers*"]

# --- Ruff (linter + formatter, замена flake8 + black + isort) ---
[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.ruff.lint.isort]
known-first-party = ["src", "workers"]

# --- Pytest ---
[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
filterwarnings = ["ignore::DeprecationWarning"]

# --- MyPy ---
[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
plugins = ["pydantic.mypy"]
